feat: 에러 메세자 추가

+++ front-end/src/utils/errorHandler.js
@@ -10,6 +10,10 @@ export const errorHandler = (error) => {
 
   const { status } = error;
 
+  if (error.config.url === API_URLS.AUTH.SIGN_IN && status === 404) {
+    return showError(ERROR_MESSAGES.AUTH_NOT_FOUND);
+  }
+
   if (error.config.url === API_URLS.USER.REISSUE && status === 401) {
     localStorage.clear();
     return showError(ERROR_MESSAGES.TOKEN_EXPIRED);
